fix duckdb split
