+++ jest.duckdb.config.ts
@@ -1,8 +1,8 @@
 module.exports = {
   ...require('./jest.config.ts'),
   roots: [
+    '<rootDir>/packages/malloy-db-duckdb/',
     '<rootDir>/test/src/databases/all/',
-    '<rootDir>/packages/malloy-db-duckdb/src/',
     '<rootDir>/test/src/databases/duckdb/',
     '<rootDir>/test/src/databases/duckdb-all/',
   ],
