+++ test/src/core/api.spec.ts
@@ -23,7 +23,7 @@
  */
 
 import {runtimeFor} from '../runtimes';
-import './util/db-jest-matchers';
+import '../util/db-jest-matchers';
 
 const runtime = runtimeFor('duckdb');
