+++ test/src/core/experimental-dialects.spec.ts
@@ -29,7 +29,7 @@ import {
   registerDialect,
 } from '@malloydata/malloy';
 import {testRuntimeFor} from '../runtimes';
-import './util/db-jest-matchers';
+import '../util/db-jest-matchers';
 import {DuckDBConnection} from '@malloydata/db-duckdb';
 
 const envDatabases = (
